# docs/swagger.yaml
basePath: /
definitions:
...
  enum.BookingStatus:
    enum:
    - pending
    - confirmed
    - cancelled
    - completed
    type: string
    x-enum-varnames:
    - Pending
    - Confirmed
    - Cancelled
    - Completed
  enum.PaymentStatus:
    enum:
    - partial
    - unpaid
    - paid
    type: string
    x-enum-varnames:
    - Partial
    - Unpaid
    - Paid
...
  models.Booking:
    description: Represents a booking with associated user, service, start date, end
      date, total price, status, and payment status.
    properties:
      created_at:
        description: |-
          Timestamp when the booking was created
          @example "2025-04-10T14:22:00Z"
        type: string
      end_date:
        description: |-
          The end date of the booking
          @example "2025-04-15T17:00:00Z"
        type: string
      id:
        description: |-
          ID of the booking, automatically incremented
          @example 1
        type: integer
      payment_status:
        allOf:
        - $ref: '#/definitions/enum.PaymentStatus'
        description: |-
          Payment status of the booking (e.g., unpaid, paid, etc.)
          @example "unpaid"
      service_id:
        description: |-
          ID of the associated service for the booking
          @example 1
        type: integer
      start_date:
        description: |-
          The start date of the booking
          @example "2025-04-15T09:00:00Z"
        type: string
      status:
        allOf:
        - $ref: '#/definitions/enum.BookingStatus'
        description: |-
          Status of the booking (e.g., pending, confirmed, etc.)
          @example "pending"
      total_price:
        description: |-
          Total price of the booking
          @example 250.75
        type: number
      updated_at:
        description: |-
          Timestamp when the booking was last updated
          @example "2025-04-10T14:22:00Z"
        type: string
      user_id:
        description: |-
          ID of the user making the booking
          @example "123e4567-e89b-12d3-a456-426614174000"
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: QFirst is a robust and easy-to-use booking system designed to help
    businesses manage appointments and reservations efficiently.
  title: QFirst API Documentation
  version: "1.0"
paths:
...
  /bookings:
    get:
      consumes:
      - application/json
      description: Retrieve a list of bookings, with optional filtering by user ID,
        service ID, or status.
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Service ID
        in: query
        name: service_id
        type: string
      - description: Booking Status
        enum:
        - pending
        - confirmed
        - cancelled
        in: query
        name: status
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Page size for pagination
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of bookings
          schema:
            items:
              $ref: '#/definitions/models.Booking'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a list of bookings.
      tags:
      - Booking
    post:
      consumes:
      - application/json
      description: Create a new booking for a user. The booking will be stored in
        the database, and the user will receive a confirmation.
      parameters:
      - description: Booking details
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/models.Booking'
      produces:
      - application/json
      responses:
        "201":
          description: Booking created successfully
          schema:
            $ref: '#/definitions/models.Booking'
        "400":
          description: Invalid input or missing data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Service or User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new booking.
      tags:
      - Booking
  /bookings/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a booking from the platform using its unique ID.
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Booking deleted successfully
        "400":
          description: Invalid Booking ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Booking not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a booking by ID.
      tags:
      - Booking
    get:
      consumes:
      - application/json
      description: Retrieve a booking by its unique ID.
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Booking found
          schema:
            $ref: '#/definitions/models.Booking'
        "400":
          description: Invalid Booking ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Booking not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a booking by ID.
      tags:
      - Booking
    put:
      consumes:
      - application/json
      description: Update an existing booking by its ID. The booking details will
        be updated, and the changes will be saved to the database.
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      - description: Booking details to update
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/models.Booking'
      produces:
      - application/json
      responses:
        "200":
          description: Booking updated successfully
          schema:
            $ref: '#/definitions/models.Booking'
        "400":
          description: Invalid input or missing data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Booking not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an existing booking.
      tags:
      - Booking
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
  CsrfToken:
    description: CSRF protection token.
    in: header
    name: X-CSRF-Token
    type: apiKey
swagger: "2.0"